+++ src/main/java/net/MrGise/mmm/block/entity/renderer/BowlBlockEntityRenderer.java
@@ -24,8 +24,7 @@ public class BowlBlockEntityRenderer implements BlockEntityRenderer<BowlBlockEnt
     private final boolean debug = false;
 
 
-    public BowlBlockEntityRenderer(BlockEntityRendererProvider.Context context) {
-    }
+    public BowlBlockEntityRenderer(BlockEntityRendererProvider.Context context) {}
 
     @Override
     public void render(BowlBlockEntity blockEntity, float partialTick, PoseStack poseStack,
@@ -59,26 +58,20 @@ public class BowlBlockEntityRenderer implements BlockEntityRenderer<BowlBlockEnt
         for (int stackIndex = 0; stackIndex < stackRenderCount; ++stackIndex) {
             poseStack.pushPose();
 
-            float stackAngleF = stackIntervalAngle * stackIndex;
+            float stackAngleF = stackIntervalAngle * stackIndex + blockEntity.getYRotOffset();
             float stackAngleR = stackAngleF * Mth.DEG_TO_RAD;
 
             float xOffset = Mth.cos(stackAngleR) * distFromCenter;
             float zOffset = -Mth.sin(stackAngleR) * distFromCenter;
 
             poseStack.translate(0.5f + xOffset, 0.15f, 0.5f + zOffset);
-            poseStack.scale(0.5f, 0.5f, 0.5f);
+            poseStack.scale(1f, 1f, 1f);
 
             poseStack.mulPose(Axis.YP.rotationDegrees(stackAngleF));
-            poseStack.mulPose(Axis.XP.rotationDegrees(180));
-            poseStack.mulPose(Axis.ZN.rotationDegrees(90));
-            poseStack.mulPose(Axis.YP.rotationDegrees(90));
-            poseStack.mulPose(Axis.XP.rotationDegrees(20));
-            poseStack.mulPose(Axis.YP.rotationDegrees(180));
-            poseStack.mulPose(Axis.XP.rotationDegrees(ClientConfig.BOWL_XROT_OFFSET.get()));
-            poseStack.mulPose(Axis.YP.rotationDegrees(ClientConfig.BOWL_YROT_OFFSET.get()));
-            poseStack.mulPose(Axis.ZP.rotationDegrees(ClientConfig.BOWL_ZROT_OFFSET.get()));
+            poseStack.mulPose(Axis.YN.rotationDegrees(90));
+            poseStack.mulPose(Axis.XN.rotationDegrees(70));
 
-            itemRenderer.renderStatic(items.get(stackIndex), ItemDisplayContext.FIXED, getLightLevel(level, blockEntity.getBlockPos()),
+            itemRenderer.renderStatic(items.get(stackIndex), ItemDisplayContext.GROUND, getLightLevel(level, blockEntity.getBlockPos()),
                     OverlayTexture.NO_OVERLAY, poseStack, bufferSource, level, 1);
 
             poseStack.popPose();