+++ src/main/java/net/MrGise/mmm/constants/Numbers.java
@@ -1,0 +1,6 @@
+package net.MrGise.mmm.constants;
+
+public class Numbers {
+    public static int CIRCLE_ANGLES = 360;
+
+}