+++ src/main/java/net/MrGise/mmm/recipe/BowlRecipe.java
@@ -4,7 +4,6 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonSyntaxException;
-import com.google.gson.annotations.Since;
 import com.simibubi.create.foundation.fluid.FluidIngredient;
 import net.MrGise.mmm.block.entity.BowlBlockEntity;
 import net.minecraft.core.NonNullList;
@@ -17,7 +16,6 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.*;
 import net.minecraft.world.level.Level;
 import net.minecraftforge.fluids.FluidStack;
-import org.checkerframework.framework.qual.Unused;
 
 import javax.annotation.Nullable;
 import java.util.ArrayList;